small bugfixes & move walletGen into src/utils
