+++ src/App.js
@@ -8,7 +8,7 @@ import DepositCard from "./components/depositCard";
 import { getConnextClient } from "connext/dist/Connext.js";
 import ProviderOptions from "./utils/ProviderOptions.ts";
 import clientProvider from "./utils/web3/clientProvider.ts";
-import { createWalletFromMnemonic } from "./walletGen";
+import { createWalletFromMnemonic, createWallet } from "./utils/walletGen";
 import { Paper, withStyles, Grid } from "@material-ui/core";
 import AppBarComponent from "./components/AppBar";
 import SettingsCard from "./components/settingsCard";
@@ -16,7 +16,6 @@ import ReceiveCard from "./components/receiveCard";
 import SendCard from "./components/sendCard";
 import CashOutCard from "./components/cashOutCard";
 import SupportCard from "./components/supportCard";
-import { createWallet } from "./walletGen";
 import RedeemCard from "./components/redeemCard";
 import SetupCard from "./components/setupCard";
 import Confirmations from "./components/Confirmations";
